requirements.txt
addonfactory-splunk-conf-parser-lib
pyyaml
